Backend/requirements.txt
django>=4.2,<5.1
djangorestframework>=3.14
django-cors-headers>=4.3
requests>=2.31
